# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: ScholarAgent API
  description: |
    ScholarAgent - 学术智能体 API

    一个基于 AI 的学术文献检索与分析智能体系统，支持多数据源检索、智能查询规划、
    证据验证和文献推荐。

    ## 主要功能

    - **智能查询规划**: 使用 LLM 将复杂学术问题分解为多个子查询
    - **多数据源检索**: 集成 Semantic Scholar、OpenAlex、CrossRef、arXiv
    - **证据验证**: 从论文全文中提取支持/反驳证据
    - **文献推荐**: 基于用户反馈的个性化推荐
  version: 1.0.0
  contact:
    name: ScholarAgent Team
    email: team@scholar-agent.dev
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: 本地开发服务器
  - url: https://api.scholar-agent.dev
    description: 生产服务器
...
paths:
...
  /api/v1/recommendations:
    post:
      operationId: getRecommendations
      summary: 获取文献推荐
      description: 基于用户历史和偏好获取个性化文献推荐
      tags: [recommendations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecommendationRequest'
      responses:
        '200':
          description: 推荐结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecommendationResponse'
  /api/v1/recommendations/feedback:
    post:
      operationId: submitFeedback
      summary: 提交推荐反馈
      description: 用户对推荐结果的反馈，用于改进推荐算法
      tags: [recommendations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecommendationFeedback'
      responses:
        '200':
          description: 反馈已记录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  # =============================================================================
  # 证据接口
  # =============================================================================
...
  /api/v1/model/config:
    post:
      operationId: saveModelConfig
      summary: 保存模型配置
      description: 保存用户的 LLM 模型配置
      tags: [model]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelConfig'
      responses:
        '200':
          description: 配置已保存
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /api/v1/model/test:
    post:
      operationId: testModelConnection
      summary: 测试模型连通性
      description: 测试 LLM 模型配置是否正确，能否正常调用
      tags: [model]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelTestRequest'
      responses:
        '200':
          description: 测试结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelTestResponse'

  # =============================================================================
  # 健康检查
  # =============================================================================
...
components:
...
  schemas:
...
    RecommendationRequest:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
          description: 用户 ID（可选）
        context:
          type: string
          description: 推荐上下文
          example: "papers similar to my recent searches on transformer models"
        limit:
          type: integer
          default: 10
          minimum: 1
          maximum: 50
    RecommendationFeedback:
      type: object
      required: [recommendation_id, feedback_type]
      properties:
        recommendation_id:
          type: string
          format: uuid
        feedback_type:
          type: string
          enum: [helpful, not_helpful, saved, dismissed]
        comment:
          type: string
          maxLength: 500
    ModelConfig:
      type: object
      required: [provider, model_name]
      properties:
        provider:
          type: string
          enum: [openai, anthropic, ollama]
          description: LLM 提供商
        model_name:
          type: string
          description: 模型名称
          example: "gpt-4o"
        api_key:
          type: string
          description: API Key（可选，使用环境变量则无需填写）
        base_url:
          type: string
          format: uri
          description: 自定义 API 地址
        temperature:
          type: number
          minimum: 0
          maximum: 2
          default: 0.7
        max_tokens:
          type: integer
          minimum: 1
          maximum: 128000
          default: 4096
    ModelTestRequest:
      type: object
      required: [provider, model_name]
      properties:
        provider:
          type: string
          enum: [openai, anthropic, ollama]
        model_name:
          type: string
        api_key:
          type: string
        base_url:
          type: string
          format: uri

    # ----- 响应 Schema -----
...
    Paper:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        authors:
          type: array
          items:
            type: string
        abstract:
          type: string
        year:
          type: integer
        venue:
          type: string
          nullable: true
        citation_count:
          type: integer
        doi:
          type: string
          nullable: true
        url:
          type: string
          format: uri
        pdf_url:
          type: string
          format: uri
          nullable: true
        source:
          $ref: '#/components/schemas/DataSource'
        relevance_score:
          type: number
          minimum: 0
          maximum: 1
        is_open_access:
          type: boolean
...
    Recommendation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        paper:
          $ref: '#/components/schemas/Paper'
        score:
          type: number
          minimum: 0
          maximum: 1
        reason:
          type: string
          description: 推荐理由
    RecommendationResponse:
      type: object
      properties:
        recommendations:
          type: array
          items:
            $ref: '#/components/schemas/Recommendation'
        has_more:
          type: boolean
...
    ModelTestResponse:
      type: object
      properties:
        success:
          type: boolean
        latency_ms:
          type: number
          description: 响应延迟（毫秒）
        model_info:
          type: object
          properties:
            provider:
              type: string
            model:
              type: string
            context_window:
              type: integer
        error:
          type: string
          nullable: true
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string

    # ----- 枚举 -----
    DataSource:
      type: string
      enum:
        - semantic_scholar
        - openalex
        - crossref
        - arxiv
      description: 学术数据源
...